UPSTREAM: PCI: qcom: Add support for configuring BDF to SID mapping for SM8250
For SM8250, we need to write the BDF to SID mapping in PCIe controller
register space for proper working. This is accomplished by extracting
the BDF and SID values from "iommu-map" property in DT and writing those
in the register address calculated from the hash value of BDF. In case
of collisions, the index of the next entry will also be written.

For the sake of it, let's introduce a "config_sid" callback and do it
conditionally for SM8250.

Link: https://lore.kernel.org/r/20201208121402.178011-4-mani@kernel.org
Signed-off-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>
Signed-off-by: Lorenzo Pieralisi <lorenzo.pieralisi@arm.com>
Reviewed-by: Bjorn Andersson <bjorn.andersson@linaro.org>
(cherry picked from commit 4c9398822106c366d88c8c68ddf44bd371d39961)

BUG=b:186590981
TEST=emerge-herobrine sys-kernel/chromeos-kernel-5_10

Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Change-Id: Ia13439b2394955e8cdfa4bc0d4f8827ff2fb796d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3049969
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>