UPSTREAM: ath11k: mhi: print a warning if firmware crashed
There was no way to detect if the firmware crashed so add a warning. At the
moment the firmware is not restarted or anything like that, so when this
happens ath11k modules need to be reloaded.

Tested-on: QCA6390 hw2.0 PCI WLAN.HST.1.0.1-01740-QCAHSTSWPLZ_V2_TO_X86-1

Signed-off-by: Kalle Valo <kvalo@codeaurora.org>
Link: https://lore.kernel.org/r/1607609124-17250-2-git-send-email-kvalo@codeaurora.org
(cherry picked from commit fc46e1b2a24a0c85e8469576f66f9a991411bfc7)

BUG=b:190968626
TEST=build passes

Signed-off-by: Abhishek Kumar <kuabhs@chromium.org>
Change-Id: I741add46aee2474c3945e649c7b2a615ed5144fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2974586
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>