misc: cs40l25: Store trigger queue in new composite structure
Take the output of the existing trigger queue parser and store this
in the new composite data structure. This is a temporary measure to
allow the rest of the code to be migrated one step at a time. Once all
the users are migrated across the parser will be replaced to fill this
structure directly.

Change-Id: Iffba20aac7adf227896698e28879e024ae304826
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>