UPSTREAM: ath11k: Fix error code in ath11k_core_suspend()
The "if (!ret)" condition is inverted and it should be "if (ret)".  It means
that we return success when we had intended to return an error code. This also
caused a spurious warning even when the suspend was successful:

[  297.186612] ath11k_pci 0000:06:00.0: failed to suspend hif: 0

Fixes: d1b0c33850d2 ("ath11k: implement suspend for QCA6390 PCI devices")
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Kalle Valo <kvalo@codeaurora.org>
Link: https://lore.kernel.org/r/X9nF17L2/EKOSbn/@mwanda
(cherry picked from commit 9b09456258ea2f35fc8a99c4ac4829dcba0ca4be)

BUG=b:190968626
TEST=build passes

Signed-off-by: Abhishek Kumar <kuabhs@chromium.org>
Change-Id: I7c4a69c88af026b792e3e6b51c5523f714068372
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2974603
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>