CHROMIUM: iwl7000: dbg: don't return -1 in iwl_dump_ini_paging_ranges()
If paging_cnt is 0 in the dram_init structure, we return -1 in
iwl_dump_ini_paging_ranges(), but the function type is u32.  This causes
us to treat the returned value as 0xffffffff, which leads to a NULL
dereference in the init flow.

Fix this by checking if paging_cnt is zero and return zero in that case.

Change-Id: I49050af1456a88a03d617ee62639d45737ea3599
Signed-off-by: Mordechay Goodstein <mordechay.goodstein@intel.com>
iwl7000-tree: e8e16ad58edbd4391986ade263642d78bf7ececd