input: cs40l26: Changes to sysfs controls
Removes all debug controls as these should not
be available to users. The exception to this is
the debug_hibernate control which will be removed
when runtime power management is added.

The dsp_state sysfs control replaces halo_state as
there is only one valid value for halo_state it is
not of great interest.

Removes the num_waves sysfs control as this value
is reported in dmesg upon inserting the kernel module.

Change-Id: I9a638f879c1aae86d4fa17776a7260b965b637f9
Signed-off-by: Fred Treven <fred.treven@cirrus.com>