firmware: cirrus: Move CL DSP to drivers/firmware
The Cirrus Logic DSP firmware driver was being used
exclusively with input haptics drivers and was located
in the drivers/input/misc directory.

Since it is a driver exclusively used for firmware
configuration it should instead reside in drivers/firmware.

This patch creates a directory for all Cirrus Logic
firmware drivers and moves the CL DSP firmware driver
into it.

Change-Id: Ib2e6c5e7362fbe22d1e2d0d0b649fc8d46aa8e08
Signed-off-by: Fred Treven <fred.treven@cirrus.com>